--- lib/colors/main_colors.dart
@@ -1,5 +1,0 @@
-import 'dart:ui';
-
-class MainColors {
-  static const Color main03 = Color(0xffFF8282);
-}
