+++ lib/designsystem/palette.dart
@@ -1,0 +1,10 @@
+import 'dart:ui';
+
+class Palette {
+  static const primary = Color(0xFFFF8282);
+  static const gray02 = Color(0xFFEFF0F5);
+  static const gray03 = Color(0xFFD7D8DC);
+  static const gray04 = Color(0xFFBFC0C4);
+  static const gray06 = Color(0xFF8F9093);
+  static const gray10 = Color(0xFF303031);
+}
