+++ lib/main.dart
@@ -4,16 +4,14 @@ import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:kakao_flutter_sdk/kakao_flutter_sdk.dart';
 
-import 'assets/designsystem/palette.dart';
+import 'designSystem/palette.dart';
 
 void main() async {
-
   WidgetsFlutterBinding.ensureInitialized();
-  await dotenv.load(fileName: "lib/.env");	// 추가
+  await dotenv.load(fileName: "lib/.env"); // 추가
 
   KakaoSdk.init(
     nativeAppKey: dotenv.env['KAKAO_APP_KEY'],
-
   );
   runApp(Birca());
 }
@@ -28,11 +26,9 @@ class Birca extends StatelessWidget {
       title: 'Birca',
       theme: ThemeData(
         primaryColor: Palette.primary,
-        textTheme: const TextTheme(
-            bodyMedium: TextStyle(color: Color(0xff303031))
-        ),
-        colorScheme: ColorScheme.fromSeed(
-            seedColor: Colors.white),
+        textTheme:
+            const TextTheme(bodyMedium: TextStyle(color: Color(0xff303031))),
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.white),
         useMaterial3: true,
       ),
       home: const Login(),
