+++ lib/view/onboarding/apply_nickname_screen.dart
@@ -2,7 +2,7 @@ import 'package:birca/widgets/appbar.dart';
 import 'package:flutter/material.dart';
 import 'package:birca/widgets/progressbar.dart';
 
-import '../../assets/designsystem/palette.dart';
+import '../../designSystem/palette.dart';
 import '../../widgets/button.dart';
 
 class ApplyNickNameScreen extends StatelessWidget {
